% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrateRun.R
\name{equilibrateRun}
\alias{equilibrateRun}
\title{Find equilibrium temperatures on January 1}
\usage{
equilibrateRun(odbcConnection, binStats, channelVolume, channelSurfaceArea,
  surfaceShade, firstBin, lastBin, timeStep, runID, internal = FALSE,
  startWith)
}
\arguments{
\item{odbcConnection}{RODBC object}

\item{binStats}{data frame returned from hyporheicBins()}

\item{channelVolume}{numeric}

\item{channelSurfaceArea}{numeric}

\item{surfaceShade}{numeric}

\item{firstBin}{numeric}

\item{lastBin}{numeric}

\item{timeStep}{numeric}

\item{runID}{character}

\item{internal}{print java output in console? logical}

\item{startWith}{values other that 1 deg C to start with}
}
\value{
data.frame of equilibrized temperatures
}
\description{
Find equilibrium temperatures on January 1
}
