% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTempTool.R
\name{runTempTool}
\alias{runTempTool}
\title{run TempTool}
\usage{
runTempTool(odbcConnection, binStats, initTemps = NULL, channelVolume,
  channelSurfaceArea, surfaceShade, firstBin, lastBin, timeStep, yearsToRun,
  outputInterval, runID = "", outputIndex = NULL, internal = TRUE)
}
\arguments{
\item{odbcConnection}{RODBC object}

\item{binStats}{data frame returned from hyporheicBins()}

\item{channelVolume}{numeric}

\item{channelSurfaceArea}{numeric}

\item{surfaceShade}{numeric}

\item{firstBin}{numeric}

\item{lastBin}{numeric}

\item{timeStep}{numeric}

\item{yearsToRun}{numeric (years)}

\item{runID}{character}

\item{outputIndex}{lubridate object}

\item{internal}{print java output? logical}
}
\value{
list of tts objects from each model cell
}
\description{
run TempTool
}
