% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setParameters.R
\name{setParameters}
\alias{setParameters}
\title{Set the model parameters.}
\usage{
setParameters(firstBin, lastBin, odbcConnection, initTemps = NULL,
  surfaceShade, channelSurfaceArea, channelVolume, binStats)
}
\arguments{
\item{firstBin}{numeric. The smallest TSZ in model run.}

\item{lastBin}{numeric. The largest TSZ in model run.}

\item{odbcConnection}{RODBC object.}

\item{initTemps}{dataframe of starting temperatures. The default is initTemps = NULL, which starts all model cells at 1 degree C.}

\item{surfaceShade}{numeric. Proportion of channel under shade.}

\item{channelSurfaceArea}{numeric. (m2)}

\item{channelVolume}{numeric. (m3)}

\item{binStats}{dataframe from hyporheicBins(). Geometry of Hyporheic Zone.}
}
\value{
Does not return anything, check model database to see changes.
}
\description{
Set the model parameters.
}
