% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSkeleton.R
\name{setSkeleton}
\alias{setSkeleton}
\title{Set the model configuration tables.}
\usage{
setSkeleton(firstBin, lastBin, odbcConnection)
}
\arguments{
\item{firstBin}{numeric, the smallest TSZ considered in model run}

\item{lastBin}{numeric, the largest TSZ in model run}

\item{odbcConnection}{RODBC object}
}
\value{
Does not return anything, check model database to see changes.
}
\description{
Set the model configuration tables.
}
