% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTiming.R
\name{setTiming}
\alias{setTiming}
\title{Set model timing.}
\usage{
setTiming(odbcConnection, timeStep, yearsToRun, outputInterval)
}
\arguments{
\item{odbcConnection}{RODBC object.}

\item{timeStep}{model time step in seconds. numeric.}

\item{yearsToRun}{how long to run model, in years. numeric.}

\item{outputInterval}{model will output at this time, in seconds. numeric.}
}
\value{
Does not return anything, check model database to see changes.
}
\description{
Set model timing.
}
