% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsz3d.R
\name{tsz3d}
\alias{tsz3d}
\title{Plot change in channel temperature per hour}
\usage{
tsz3d(TSZnum, dates, modelTemps, channelDepth, channelSurfaceArea, binStats,
  colors, numColors, colorGroupings, zlim, phi = 30, theta = 30,
  zeroGrid = TRUE)
}
\arguments{
\item{TSZnum}{the number of TSZs in model. numeric.}

\item{dates}{the date or dates to plot. lubridate object.}

\item{modelTemps}{list of model temperatures. tts object.}

\item{channelDepth}{numeric.}

\item{channelSurfaceArea}{numeric.}

\item{binStats}{TSZ geometry. dataframe output from hyporheicBins()}

\item{colors}{color ramp. numeric vector.}

\item{numColors}{the number of colors in colors vector. numeric.}

\item{colorGroupings}{the number of divisions in colors. numeric.}

\item{zlim}{z-axis upper and lower bounds. numeric vector of length 2.}

\item{phi}{phi}

\item{theta}{theta}

\item{zeroGrid}{plot a  zero plane? TRUE/FALSE. default TRUE.}
}
\value{
3-D surface plot of change in channel temperature per hour contributed by each TSZ.
}
\description{
Plot change in channel temperature per hour
}
